Restructure database for attestations
