--- database/db_models.py
@@ -1,2 +1,0 @@
-from database import db
-from .notification_models import *  # NOQA
