+++ database/models.py
@@ -1,6 +1,24 @@
 from database import db
-from sqlalchemy import func
+from datetime import datetime
 from enum import Enum
+from sqlalchemy import func
+
+
+class AttestationTypes(Enum):
+    PHONE = 1
+    EMAIL = 2
+    AIRBNB = 3
+    FACEBOOK = 4
+    TWITTER = 5
+
+
+class Attestation(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    method = db.Column(db.Enum(AttestationTypes))
+    eth_address = db.Column(db.String)
+    value = db.Column(db.String)
+    signature = db.Column(db.String)
+    created_at = db.Column(db.DateTime, default=datetime.utcnow)
 
 
 class EthNotificationTypes(Enum):
@@ -15,8 +33,6 @@ class EthNotificationEndpoint(db.Model):
     type = db.Column(db.Enum(EthNotificationTypes))
     active = db.Column(db.Boolean())
     verified = db.Column(db.Boolean())
-    created_at = db.Column(
-        db.DateTime(
-            timezone=True),
-        server_default=func.now())
+    created_at = db.Column(db.DateTime, default=datetime.utcnow)
     expires_at = db.Column(db.DateTime(timezone=True))
+
