+++ pytest.ini
@@ -11,6 +11,4 @@ env =
   TWITTER_CONSUMER_KEY=twitter-consumer-key
   TWITTER_CONSUMER_SECRET=twitter-consumer-secret
   ATTESTATION_SIGNING_KEY=0x0000000000000000000000000000000000000000000000000000000000000001
-flakes-ignore =
-    database/db_models.py UnusedImport ImportStarUsed
 codestyle_max_line_length = 100
