+++ tests/services/test_notifier.py
@@ -1,4 +1,4 @@
-from database.db_models import EthNotificationEndpoint, EthNotificationTypes
+from database.models import EthNotificationEndpoint, EthNotificationTypes
 from logic.notifier_service import register_eth_notification
 from tests.helpers.eth_utils import sample_eth_address, str_eth
